+++ src/cmd/jec/CreateBootstrap.ts
@@ -1,0 +1,61 @@
+//  DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
+//
+//   Copyright 2016-2018 Pascal ECHEMANN.
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License.
+
+import {Command} from "../Command";
+import {CliLogger} from "../../utils/CliLogger";
+import {TemplateBuilder} from "../../template/util/TemplateBuilder";
+import {BootstrapTemplateGenerator} from "../../template/generator/BootstrapTemplateGenerator";
+import * as fs from "fs";
+import * as path from "path";
+
+/**
+ * The <code>CreateBootstrap</code> command allows create a custom bootstrap
+ * component.
+ */
+export class CreateBootstrap implements Command {
+  
+  //////////////////////////////////////////////////////////////////////////////
+  // Constructor function
+  //////////////////////////////////////////////////////////////////////////////
+
+  /**
+   * Creates a new <code>CreateBootstrap</code> instance.
+   */
+  constructor() {}
+
+  //////////////////////////////////////////////////////////////////////////////
+  // Public methods
+  //////////////////////////////////////////////////////////////////////////////
+
+  /**
+   * @inheritDoc
+   */
+  public run(argv:any):void {
+    const logger:CliLogger= CliLogger.getInstance();
+    const builder:TemplateBuilder = new TemplateBuilder();
+    const name:string = argv.name;
+    const template:string = builder.build(BootstrapTemplateGenerator, argv);
+    fs.writeFile(`${name}.ts`, template, (err:NodeJS.ErrnoException | null) => {
+      if(err) {
+        logger.error(err);
+      } else {
+        logger.log(
+          `Bootstrap file with name '${name}.ts' created in '${process.cwd()}'.`
+        );
+      }
+    });
+  }
+}