+++ CHANGELOG.md
@@ -10,7 +10,8 @@
 - `jec` package refactoring
 - Adding the `jec version` command
 - Adding the `jec glasscat-version` command
-- Adding the `help` commands
+- Adding the `help` command
+- Adding the `jec create-bootstrap` command
 
 <a name="jec-cli-0.1.1"></a>
 ## **0.1.1** (2018-03-18)